Former-commit-id: df5bca8f22523ff5174c0abf8b9735bf066fb662 [formerly fae4b9d3b31f4f47870d58322b0044c6c427b76a] [formerly da11148a1c75b423ef5b3ecfb874dcbd0b7c2efb [formerly 2d0d2f85a1bf02b5033c98d39abaffc406072782]]
Former-commit-id: 86c9d5384e1cd070cad4df9cf535b75aa5f8a019 [formerly 54b637a10e68f5d6d52ac08f0495d3f5c7b8211b]
Former-commit-id: d8f65cdcd654e9d631b8e62af42dfdd2c5e3415d

+++ Frame/SDK/Core/Base/AFLockFreeQueue.h
@@ -23,7 +23,6 @@
 #include "SDK/Core/Base/AFPlatform.hpp"
 #include "common/readerwriterqueue.h"
 
-
 template<typename T>
 class AFLockFreeQueue
 {
@@ -47,11 +46,11 @@ public:
         return mList.try_dequeue(object);
     }
 
-    int Count()
+    size_t Count()
     {
         return mList.size_approx();
     }
 
 private:
-    moodycamel::BlockingReaderWriterQueue <T> mList;
+    moodycamel::BlockingReaderWriterQueue<T> mList;
 };
